# the-nastar-server/apispecs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "nastar api server",
    "description": "API for online store rest api",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:3000/api/docs", "description": "Local server" }],
  "paths": {
...
    "/products": {
      "get": {
        "responses": {
          "200": {
            "description": "products list"
          }
        }
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  }
                },
                "required": ["name", "description", "price"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "product created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "result": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          }
                        },
                        "required": ["id", "name", "description", "price"]
                      }
                    }
                  },
                  "required": ["id", "username", "password"]
                }
              }
            }
          },
          "401": {
            "description": "method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["id", "username", "password"]
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "responses": {
          "200": {
            "description": "product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    }
                  },
                  "required": ["id", "name", "description", "price"]
                }
              }
            }
          }
        }
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  }
                },
                "required": ["name", "description", "price"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "price": {
                          "type": "number"
                        }
                      },
                      "required": ["id", "name", "description", "price"]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "responses": {
          "200": {
            "description": "product deleted"
          }
        }
      }
    },
...
  }
}